pyproject.toml
# PyProject configuration for mammography-pipelines
# Educational Research Project - NOT for clinical use

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mammography-pipelines"
version = "1.0.0"
description = "Mammography pipelines for breast density analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Research Team", email = "research@example.com"},
]
keywords = ["medical-imaging", "deep-learning", "breast-density", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "pydicom>=2.4.0",
    "Pillow>=9.5.0",
    "nibabel>=5.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pandera>=0.20.1",
    "scipy>=1.10.0",
    "umap-learn>=0.5.3",
    "hdbscan>=0.8.29",
    "plotly>=5.15.0",
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
    "optuna>=3.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.41.0",
    "seaborn>=0.12.0",
    "click>=8.1.0",
    "rich>=13.5.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "nbsphinx>=0.9.0",
]

[project.scripts]
mammography = "mammography.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B006",  # do not use mutable data structures for argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "E203",  # whitespace before ':'
    "E402",  # module level import not at top of file
    "E731",  # do not assign a lambda expression, use a def
    "TID252", # relative imports are banned
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "TID252"]

[tool.ruff.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | archive
  | data/raw
  | results
  | logs
)/
'''

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true
color_output = true
error_summary = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "torchaudio.*",
    "pydicom.*",
    "PIL.*",
    "nibabel.*",
    "sklearn.*",
    "umap.*",
    "hdbscan.*",
    "plotly.*",
    "mlflow.*",
    "wandb.*",
    "hydra.*",
    "omegaconf.*",
    "pydantic.*",
    "tqdm.*",
    "matplotlib.*",
    "seaborn.*",
    "pytest.*",
    "black.*",
    "ruff.*",
    "click.*",
    "rich.*",
    "typer.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--cov=mammography.analysis",
    "--cov=mammography.cli",
    "--cov=mammography.data",
    "--cov=mammography.models.nets",
    "--cov=mammography.training",
    "--cov=mammography.io.dicom",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=50",
    "--durations=10",
    "--maxfail=5",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "slow: Slow running tests",
    "gpu: Tests requiring GPU",
    "cpu: Tests that can run on CPU",
    "preprocessing: Tests for preprocessing modules",
    "embedding: Tests for embedding extraction",
    "clustering: Tests for clustering algorithms",
    "semi_supervised: Tests for semi-supervised learning",
    "baseline: Tests for supervised baseline",
    "medical_data: Tests involving medical data",
    "reproducibility: Tests for reproducibility",
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::FutureWarning",
    "ignore::RuntimeWarning:pydicom.*",
]
# env requires pytest-env plugin, commenting out for now
# env = [
#     "PYTHONPATH = src",
#     "PYTHONHASHSEED = 42",
#     "CUDA_VISIBLE_DEVICES = 0",
#     "OMP_NUM_THREADS = 1",
#     "MKL_NUM_THREADS = 1",
# ]
minversion = "7.0"
# timeout = 300  # requires pytest-timeout plugin
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
# collect_ignore is deprecated in pytest 8.x, use norecursedirs or collect_ignore_glob instead
# collect_ignore = [
#     "setup.py",
#     "docs/",
#     "archive/",
#     "data/raw/",
#     "results/",
#     "logs/",
# ]
norecursedirs = [
    "setup.py",
    "docs",
    "archive",
    "data/raw",
    "results",
    "logs",
]
junit_family = "xunit2"
junit_logging = "all"
junit_log_passing_tests = true
